--- a/src/cache_element.jl
+++ b/src/cache_element.jl
@@ -20,7 +20,7 @@ end
 
 @inline isdirty(element::CacheElement) = element.dirty
 
-# The reason to do this using a macro instead of using something like get!(element::CacheElement, callable)
+# The reason for doing this using a macro instead of using something like get!(element::CacheElement, callable)
 # is that in all use cases so far, the callable would be a closure over some non-isbits types, meaning that
 # creating the closure allocates.
 # Example showing this allocation behavior: @benchmark (fun = () -> a.x) setup = a = Ref(rand(Int64))