--- a/src/mechanism_state.jl
+++ b/src/mechanism_state.jl
@@ -130,14 +130,14 @@ velocity(state::MechanismState, joint::Joint) = edge_to_parent_data(state_vertex
 non_root_vertices(state::MechanismState) = state.nonRootTopoSortedStateVertices
 
 function setdirty!(state::MechanismState)
-    for vertex in state.nonRootTopoSortedStateVertices
+    for vertex in non_root_vertices(state)
         setdirty!(vertex_data(vertex))
         setdirty!(edge_to_parent_data(vertex))
     end
 end
 
 function zero_configuration!(state::MechanismState)
-    for vertex in state.nonRootTopoSortedStateVertices
+    for vertex in non_root_vertices(state)
         zero_configuration!(edge_to_parent_data(vertex))
     end
     setdirty!(state)
@@ -152,7 +152,7 @@ end
 zero!(state::MechanismState) = begin zero_configuration!(state); zero_velocity!(state) end
 
 function rand_configuration!(state::MechanismState)
-    for vertex in state.nonRootTopoSortedStateVertices
+    for vertex in non_root_vertices(state)
         rand_configuration!(edge_to_parent_data(vertex))
     end
     setdirty!(state)
@@ -256,7 +256,7 @@ momentum{X, M, C}(vertex::TreeVertex{RigidBodyState{M, C}, JointState{X, M, C}})
 momentum_rate_bias{X, M, C}(vertex::TreeVertex{RigidBodyState{M, C}, JointState{X, M, C}}) = newton_euler(vertex, bias_acceleration(vertex))
 
 function configuration_derivative!{X}(out::AbstractVector{X}, state::MechanismState{X})
-    for vertex in state.nonRootTopoSortedStateVertices
+    for vertex in non_root_vertices(state)
         jointState = edge_to_parent_data(vertex)
         q = configuration(jointState)
         v = velocity(jointState)