* finally got rid of all allocations in `mass_matrix!` (with https://github.com/JuliaArrays/StaticArrays.jl/pull/82, which helped a great deal).
* allocations in `dynamics!` and `inverse_dynamics!` are close to zero as well now; almost there.
* fixes https://github.com/tkoolen/RigidBodyDynamics.jl/issues/109
* some other minor performance improvements.
* simplified rtti macro a bit; it now expects a tuple.

Before (with https://github.com/JuliaArrays/StaticArrays.jl/pull/82):
```
inverse_dynamics:
BenchmarkTools.Trial: 
  memory estimate:  11.63 kb
  allocs estimate:  620
  --------------
  minimum time:     50.290 μs (0.00% GC)
  median time:      51.143 μs (0.00% GC)
  mean time:        52.789 μs (1.77% GC)
  maximum time:     1.985 ms (95.97% GC)

dynamics:
BenchmarkTools.Trial: 
  memory estimate:  16.13 kb
  allocs estimate:  869
  --------------
  minimum time:     91.392 μs (0.00% GC)
  median time:      141.795 μs (0.00% GC)
  mean time:        139.169 μs (1.19% GC)
  maximum time:     2.702 ms (93.11% GC)

mass_matrix:
BenchmarkTools.Trial: 
  memory estimate:  6.69 kb
  allocs estimate:  428
  --------------
  minimum time:     44.550 μs (0.00% GC)
  median time:      45.467 μs (0.00% GC)
  mean time:        46.422 μs (0.46% GC)
  maximum time:     1.123 ms (94.81% GC)
```

After:
```
inverse_dynamics:
BenchmarkTools.Trial: 
  memory estimate:  2.91 kb
  allocs estimate:  62
  --------------
  minimum time:     42.761 μs (0.00% GC)
  median time:      43.366 μs (0.00% GC)
  mean time:        44.233 μs (0.46% GC)
  maximum time:     2.090 ms (96.62% GC)

dynamics:
BenchmarkTools.Trial: 
  memory estimate:  2.95 kb
  allocs estimate:  64
  --------------
  minimum time:     73.487 μs (0.00% GC)
  median time:      74.677 μs (0.00% GC)
  mean time:        74.885 μs (0.00% GC)
  maximum time:     261.432 μs (0.00% GC)
 
mass_matrix:
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     34.701 μs (0.00% GC)
  median time:      34.974 μs (0.00% GC)
  mean time:        35.353 μs (0.00% GC)
  maximum time:     72.627 μs (0.00% GC)
```